requirements.txt
streamlit
numpy
pandas
scikit-learn
tensorflow
Pillow
joblib
